+++ src/django_display_ids/contrib/rest_framework/views.py
@@ -7,8 +7,6 @@ from typing import TYPE_CHECKING, Any
 from django_display_ids.conf import (
     NOT_SET,
     get_setting,
-    get_slug_field,
-    get_uuid_field,
 )
 from django_display_ids.encoding import PREFIX_PATTERN
 from django_display_ids.exceptions import (
@@ -77,12 +75,6 @@ class DisplayIDMixin:
     # These may be provided by parent classes
     kwargs: dict[str, Any]
     request: Any
-
-    def _get_uuid_field(self) -> str:
-        return get_uuid_field(self.uuid_field)
-
-    def _get_slug_field(self) -> str:
-        return get_slug_field(self.slug_field)
 
     def _get_strategies(self) -> tuple[StrategyName, ...]:
         if self.lookup_strategies is not None:
@@ -155,8 +147,8 @@ class DisplayIDMixin:
                 value=str(value),
                 strategies=self._get_strategies(),
                 prefix=self._get_display_id_prefix(model),
-                uuid_field=self._get_uuid_field(),
-                slug_field=self._get_slug_field(),
+                uuid_field=self.uuid_field,
+                slug_field=self.slug_field,
                 queryset=queryset,
             )
         except ObjectNotFoundError as e: